mcp-agent/requirements.txt
modelscope>=1.14.0
transformers>=4.36.0
accelerate>=0.25.0
datasets>=2.14.0
addict>=2.4.0
fastapi>=0.104.0
sse-starlette>=1.8.0
pydantic>=2.5.0
mcp>=0.1.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
httpx>=0.25.0
pyyaml>=6.0.0
asyncio
typing-extensions>=4.8.0